+++ common/include/format_server_node.h
@@ -8,7 +8,7 @@
 struct node_update_ret_payload {
 	int32_t pid;
 	int32_t port;
-	int32_t label;
+	int8_t label;
 };
 
 enum notify_type {
