+++ common/include/id.h
@@ -2,4 +2,4 @@
 
 #include <stdint.h>
 
-uint32_t id_generate(void);
+uint16_t id_generate(void);
