+++ common/src/routing.c
@@ -5,7 +5,7 @@
 
 #include "custom_logger.h"
 
-int32_t routing_table_fill_default(routing_table_t* table) {
+int8_t routing_table_fill_default(routing_table_t* table) {
 	size_t i;
 
 	for (i = 0; i < (size_t) NODE_COUNT; i++) {
@@ -18,7 +18,7 @@ int32_t routing_table_fill_default(routing_table_t* table) {
 	return 0;
 }
 
-int32_t routing_next_label(const routing_table_t* table, int32_t dest_label) {
+int8_t routing_next_label(const routing_table_t* table, int8_t dest_label) {
 	if (dest_label < NODE_COUNT) {
 		return table->nodes[dest_label].label;
 	} else {
@@ -26,7 +26,7 @@ int32_t routing_next_label(const routing_table_t* table, int32_t dest_label) {
 	}
 }
 
-routing_node_t routing_get(const routing_table_t* table, int32_t dest_label) {
+routing_node_t routing_get(const routing_table_t* table, int8_t dest_label) {
 	if (dest_label < NODE_COUNT) {
 		return table->nodes[dest_label];
 	} else {
@@ -38,14 +38,14 @@ routing_node_t routing_get(const routing_table_t* table, int32_t dest_label) {
 	}
 }
 
-void routing_set_label(routing_table_t* table, int32_t dest_label, int32_t next_label, int32_t metric) { // NOLINT
+void routing_set_label(routing_table_t* table, int8_t dest_label, int8_t next_label, int8_t metric) { // NOLINT
 	if (dest_label < NODE_COUNT) {
 		table->nodes[dest_label].label = next_label;
 		table->nodes[dest_label].metric = metric;
 	}
 }
 
-void routing_del(routing_table_t* table, int32_t dest_label) {
+void routing_del(routing_table_t* table, int8_t dest_label) {
 	if (dest_label < NODE_COUNT) {
 		table->nodes[dest_label].label = -1;
 		table->nodes[dest_label].metric = 0;
